requirements.txt
pytesseract==0.3.10
Pillow==9.0.1
pandas